split backtrace into its own file
